requirements.txt
crewai
langchain
langchain-groq
litellm
groq
python-dotenv
streamlit
duckduckgo-search